+++ st-wl.c
@@ -37,6 +37,8 @@
 
 char *argv0;
 
+#include "st-wl.h"
+
 #if   defined(__linux)
  #include <pty.h>
 #elif defined(__OpenBSD__) || defined(__NetBSD__) || defined(__APPLE__)
@@ -48,8 +50,6 @@ char *argv0;
 
 /* Arbitrary sizes */
 #define UTF_INVALID   0xFFFD
-#define UTF_SIZ       4
-#define ESC_BUF_SIZ   (128*UTF_SIZ)
 #define ESC_ARG_SIZ   16
 #define STR_BUF_SIZ   ESC_BUF_SIZ
 #define STR_ARG_SIZ   ESC_ARG_SIZ
@@ -58,59 +58,24 @@ char *argv0;
 #define XK_NO_MOD     0
 #define XK_SWITCH_MOD (1<<13)
 
-#define MOD_MASK_ANY	UINT_MAX
-#define MOD_MASK_NONE	0
-#define MOD_MASK_CTRL	(1<<0)
-#define MOD_MASK_ALT	(1<<1)
-#define MOD_MASK_SHIFT	(1<<2)
-#define MOD_MASK_LOGO	(1<<3)
-
-#define AXIS_VERTICAL	WL_POINTER_AXIS_VERTICAL_SCROLL
 #define AXIS_HORIZONTAL	WL_POINTER_AXIS_HORIZONTAL_SCROLL
 
 /* macros */
-#define MIN(a, b)		((a) < (b) ? (a) : (b))
-#define MAX(a, b)		((a) < (b) ? (b) : (a))
-#define LEN(a)			(sizeof(a) / sizeof(a)[0])
 #define DEFAULT(a, b)		(a) = (a) ? (a) : (b)
-#define BETWEEN(x, a, b)	((a) <= (x) && (x) <= (b))
-#define DIVCEIL(n, d)		(((n) + ((d) - 1)) / (d))
 #define ISCONTROLC0(c)		(BETWEEN(c, 0, 0x1f) || (c) == '\177')
 #define ISCONTROLC1(c)		(BETWEEN(c, 0x80, 0x9f))
 #define ISCONTROL(c)		(ISCONTROLC0(c) || ISCONTROLC1(c))
 #define ISDELIM(u)		(utf8strchr(worddelimiters, u) != NULL)
-#define LIMIT(x, a, b)		(x) = (x) < (a) ? (a) : (x) > (b) ? (b) : (x)
 #define ATTRCMP(a, b)		((a).mode != (b).mode || (a).fg != (b).fg || \
 				(a).bg != (b).bg)
-#define IS_SET(flag)		((term.mode & (flag)) != 0)
-#define TIMEDIFF(t1, t2)	((t1.tv_sec-t2.tv_sec)*1000 + \
-				(t1.tv_nsec-t2.tv_nsec)/1E6)
-#define MODBIT(x, set, bit)	((set) ? ((x) |= (bit)) : ((x) &= ~(bit)))
 
 #define TRUECOLOR(r,g,b)	(1 << 24 | (r) << 16 | (g) << 8 | (b))
-#define IS_TRUECOL(x)		(1 << 24 & (x))
 #define TRUERED(x)		(((x) & 0xff0000) >> 8)
 #define TRUEGREEN(x)		(((x) & 0xff00))
 #define TRUEBLUE(x)		(((x) & 0xff) << 8)
 
 /* constants */
 #define ISO14755CMD		"dmenu -p codepoint: </dev/null"
-
-enum glyph_attribute {
-	ATTR_NULL       = 0,
-	ATTR_BOLD       = 1 << 0,
-	ATTR_FAINT      = 1 << 1,
-	ATTR_ITALIC     = 1 << 2,
-	ATTR_UNDERLINE  = 1 << 3,
-	ATTR_BLINK      = 1 << 4,
-	ATTR_REVERSE    = 1 << 5,
-	ATTR_INVISIBLE  = 1 << 6,
-	ATTR_STRUCK     = 1 << 7,
-	ATTR_WRAP       = 1 << 8,
-	ATTR_WIDE       = 1 << 9,
-	ATTR_WDUMMY     = 1 << 10,
-	ATTR_BOLD_FAINT = ATTR_BOLD | ATTR_FAINT,
-};
 
 enum cursor_movement {
 	CURSOR_SAVE,
@@ -121,34 +86,6 @@ enum cursor_state {
 	CURSOR_DEFAULT  = 0,
 	CURSOR_WRAPNEXT = 1,
 	CURSOR_ORIGIN   = 2
-};
-
-enum term_mode {
-	MODE_WRAP        = 1 << 0,
-	MODE_INSERT      = 1 << 1,
-	MODE_APPKEYPAD   = 1 << 2,
-	MODE_ALTSCREEN   = 1 << 3,
-	MODE_CRLF        = 1 << 4,
-	MODE_MOUSEBTN    = 1 << 5,
-	MODE_MOUSEMOTION = 1 << 6,
-	MODE_REVERSE     = 1 << 7,
-	MODE_KBDLOCK     = 1 << 8,
-	MODE_HIDE        = 1 << 9,
-	MODE_ECHO        = 1 << 10,
-	MODE_APPCURSOR   = 1 << 11,
-	MODE_MOUSESGR    = 1 << 12,
-	MODE_8BIT        = 1 << 13,
-	MODE_BLINK       = 1 << 14,
-	MODE_FBLINK      = 1 << 15,
-	MODE_FOCUS       = 1 << 16,
-	MODE_MOUSEX10    = 1 << 17,
-	MODE_MOUSEMANY   = 1 << 18,
-	MODE_BRCKTPASTE  = 1 << 19,
-	MODE_PRINT       = 1 << 20,
-	MODE_UTF8        = 1 << 21,
-	MODE_SIXEL       = 1 << 22,
-	MODE_MOUSE       = MODE_MOUSEBTN|MODE_MOUSEMOTION|MODE_MOUSEX10\
-	                  |MODE_MOUSEMANY,
 };
 
 enum charset {
@@ -172,50 +109,6 @@ enum escape_state {
 	ESC_DCS        =128,
 };
 
-enum window_state {
-	WIN_VISIBLE = 1,
-	WIN_FOCUSED = 2
-};
-
-enum selection_mode {
-	SEL_IDLE = 0,
-	SEL_EMPTY = 1,
-	SEL_READY = 2
-};
-
-enum selection_type {
-	SEL_REGULAR = 1,
-	SEL_RECTANGULAR = 2
-};
-
-enum selection_snap {
-	SNAP_WORD = 1,
-	SNAP_LINE = 2
-};
-
-typedef unsigned char uchar;
-typedef unsigned int uint;
-typedef unsigned long ulong;
-typedef unsigned short ushort;
-
-typedef uint_least32_t Rune;
-
-typedef struct {
-	Rune u;           /* character code */
-	ushort mode;      /* attribute flags */
-	uint32_t fg;      /* foreground  */
-	uint32_t bg;      /* background  */
-} Glyph;
-
-typedef Glyph *Line;
-
-typedef struct {
-	Glyph attr; /* current char attributes */
-	int x;
-	int y;
-	char state;
-} TCursor;
-
 /* CSI Escape sequence structs */
 /* ESC '[' [[ [<priv>] <arg> [;]] <mode> [<mode>]] */
 typedef struct {
@@ -237,185 +130,16 @@ typedef struct {
 	int narg;              /* nb of args */
 } STREscape;
 
-/* Internal representation of the screen */
-typedef struct {
-	int row;      /* nb row */
-	int col;      /* nb col */
-	Line *line;   /* screen */
-	Line *alt;    /* alternate screen */
-	int *dirty;  /* dirtyness of lines */
-	TCursor c;    /* cursor */
-	int top;      /* top    scroll limit */
-	int bot;      /* bottom scroll limit */
-	int mode;     /* terminal mode flags */
-	int esc;      /* escape state flags */
-	char trantbl[4]; /* charset table translation */
-	int charset;  /* current charset */
-	int icharset; /* selected charset for sequence */
-	int numlock; /* lock numbers in keyboard */
-	int *tabs;
-} Term;
 
-typedef struct {
-	struct xkb_context *ctx;
-	struct xkb_keymap *keymap;
-	struct xkb_state *state;
-	xkb_mod_index_t ctrl, alt, shift, logo;
-	unsigned int mods;
-} XKB;
-
-typedef struct {
-	struct wl_display *dpy;
-	struct wl_compositor *cmp;
-	struct wl_shm *shm;
-	struct wl_seat *seat;
-	struct wl_keyboard *keyboard;
-	struct wl_pointer *pointer;
-	struct wl_data_device_manager *datadevmanager;
-	struct wl_data_device *datadev;
-	struct wl_data_offer *seloffer;
-	struct wl_surface *surface;
-	struct wl_buffer *buffer;
-	struct xdg_wm_base *wm;
-	struct xdg_surface *xdgsurface;
-	struct xdg_toplevel *xdgtoplevel;
-	XKB xkb;
-	bool configured;
-	int px, py; /* pointer x and y */
-	int tw, th; /* tty width and height */
-	int w, h; /* window width and height */
-	int ch; /* char height */
-	int cw; /* char width  */
-	int vis;
-	char state; /* focus, redraw, visible */
-	int cursor; /* cursor style */
-	struct wl_callback * framecb;
-} Wayland;
-
-typedef struct {
-	struct wld_context *ctx;
-	struct wld_font_context *fontctx;
-	struct wld_renderer *renderer;
-	struct wld_buffer *buffer, *oldbuffer;
-} WLD;
-
-typedef struct {
-	struct wl_cursor_theme *theme;
-	struct wl_cursor *cursor;
-	struct wl_surface *surface;
-} Cursor;
-
-typedef struct {
-	uint b;
-	uint mask;
-	char *s;
-} MouseShortcut;
-
-typedef struct {
-	int axis;
-	int dir;
-	uint mask;
-	char s[ESC_BUF_SIZ];
-} Axiskey;
-
-typedef struct {
-	xkb_keysym_t k;
-	uint mask;
-	char *s;
-	/* three valued logic variables: 0 indifferent, 1 on, -1 off */
-	signed char appkey;    /* application keypad */
-	signed char appcursor; /* application cursor */
-	signed char crlf;      /* crlf mode          */
-} Key;
-
-typedef struct {
-	int mode;
-	int type;
-	int snap;
-	/*
-	 * Selection variables:
-	 * nb – normalized coordinates of the beginning of the selection
-	 * ne – normalized coordinates of the end of the selection
-	 * ob – original coordinates of the beginning of the selection
-	 * oe – original coordinates of the end of the selection
-	 */
-	struct {
-		int x, y;
-	} nb, ne, ob, oe;
-
-	char *primary;
-	struct wl_data_source *source;
-	int alt;
-	uint32_t tclick1, tclick2;
-} Selection;
-
-typedef union {
-	int i;
-	uint ui;
-	float f;
-	const void *v;
-} Arg;
-
-typedef struct {
-	uint mod;
-	xkb_keysym_t keysym;
-	void (*func)(const Arg *);
-	const Arg arg;
-} Shortcut;
-
-typedef struct {
-	char str[32];
-	uint32_t key;
-	int len;
-	bool started;
-	struct timespec last;
-} Repeat;
-
-/* function definitions used in config.h */
-static void numlock(const Arg *);
-static void selpaste(const Arg *);
-static void wlzoom(const Arg *);
-static void wlzoomabs(const Arg *);
-static void wlzoomreset(const Arg *);
-static void printsel(const Arg *);
-static void printscreen(const Arg *) ;
-static void iso14755(const Arg *);
-static void toggleprinter(const Arg *);
-static void sendbreak(const Arg *);
 
 /* Config.h for applying patches and the configuration. */
 #include "config.h"
 
-/* Font structure */
-typedef struct {
-	int height;
-	int width;
-	int ascent;
-	int descent;
-	int badslant;
-	int badweight;
-	short lbearing;
-	short rbearing;
-	struct wld_font *match;
-	FcFontSet *set;
-	FcPattern *pattern;
-} Font;
-
-/* Drawing Context */
-typedef struct {
-	uint32_t col[MAX(LEN(colorname), 256)];
-	Font font, bfont, ifont, ibfont;
-} DC;
-
-static void die(const char *, ...);
-static void draw(void);
 static void redraw(void);
 static void drawregion(int, int, int, int);
 static void execsh(void);
 static void stty(void);
 static void sigchld(int);
-static void run(void);
-static void cresize(int, int);
 
 static void csidump(void);
 static void csihandle(void);
@@ -427,7 +151,6 @@ static void strhandle(void);
 static void strparse(void);
 static void strreset(void);
 
-static int tattrset(int);
 static void tprinter(char *, size_t);
 static void tdumpsel(void);
 static void tdumpline(int);
@@ -441,8 +164,6 @@ static void tinsertblankline(int);
 static int tlinelen(int);
 static void tmoveto(int, int);
 static void tmoveato(int, int);
-static void tnew(int, int);
-static void tnewline(int);
 static void tputtab(int);
 static void tputc(Rune);
 static void treset(void);
@@ -453,8 +174,6 @@ static void tsetattr(int *, int);
 static void tsetchar(Rune, Glyph *, int, int);
 static void tsetscroll(int, int);
 static void tswapscreen(void);
-static void tsetdirt(int, int);
-static void tsetdirtattr(int);
 static void tsetmode(int, int, int *, int);
 static void tfulldirt(void);
 static void techo(Rune);
@@ -463,172 +182,65 @@ static void tdectest(char );
 static void tdefutf8(char);
 static int32_t tdefcolor(int *, int *, int);
 static void tdeftran(char);
-static inline int match(uint, uint);
-static void ttynew(void);
-static size_t ttyread(void);
-static void ttyresize(void);
-static void ttysend(char *, size_t);
-static void ttywrite(const char *, size_t);
 static void tstrsequence(uchar);
 
-static inline uchar sixd_to_8bit(int);
-static void wldraws(char *, Glyph, int, int, int, int);
-static void wldrawglyph(Glyph, int, int);
-static void wlclear(int, int, int, int);
-static void wldrawcursor(void);
-static void wlinit(void);
-static void wlloadcols(void);
-static int wlsetcolorname(int, const char *);
-static void wlloadcursor(void);
-static int wlloadfont(Font *, FcPattern *);
-static void wlloadfonts(char *, double);
-static void wlsettitle(char *);
-static void wlresettitle(void);
 static void wlseturgency(int);
 static void wlsetsel(char*, uint32_t);
-static void wlunloadfont(Font *f);
-static void wlunloadfonts(void);
-static void wlresize(int, int);
 
-static void regglobal(void *, struct wl_registry *, uint32_t, const char *,
-		uint32_t);
-static void regglobalremove(void *, struct wl_registry *, uint32_t);
-static void surfenter(void *, struct wl_surface *, struct wl_output *);
-static void surfleave(void *, struct wl_surface *, struct wl_output *);
-static void framedone(void *, struct wl_callback *, uint32_t);
-static void kbdkeymap(void *, struct wl_keyboard *, uint32_t, int32_t, uint32_t);
-static void kbdenter(void *, struct wl_keyboard *, uint32_t,
-		struct wl_surface *, struct wl_array *);
-static void kbdleave(void *, struct wl_keyboard *, uint32_t,
-		struct wl_surface *);
-static void kbdkey(void *, struct wl_keyboard *, uint32_t, uint32_t, uint32_t,
-		uint32_t);
-static void kbdmodifiers(void *, struct wl_keyboard *, uint32_t, uint32_t,
-		uint32_t, uint32_t, uint32_t);
-static void kbdrepeatinfo(void *, struct wl_keyboard *, int32_t, int32_t);
-static void ptrenter(void *, struct wl_pointer *, uint32_t, struct wl_surface *,
-		wl_fixed_t, wl_fixed_t);
-static void ptrleave(void *, struct wl_pointer *, uint32_t,
-		struct wl_surface *);
-static void ptrmotion(void *, struct wl_pointer *, uint32_t,
-		wl_fixed_t, wl_fixed_t);
-static void ptrbutton(void *, struct wl_pointer *, uint32_t, uint32_t,
-		uint32_t, uint32_t);
-static void ptraxis(void *, struct wl_pointer *, uint32_t, uint32_t,
-		wl_fixed_t);
-static void wmping(void *, struct xdg_wm_base *, uint32_t);
-static void xdgsurfconfigure(void *, struct xdg_surface *, uint32_t);
-static void xdgtoplevelconfigure(void *, struct xdg_toplevel *,
-		int32_t, int32_t, struct wl_array *);
-static void xdgtoplevelclose(void *, struct xdg_toplevel *);
-static void datadevoffer(void *, struct wl_data_device *,
-		struct wl_data_offer *);
-static void datadeventer(void *, struct wl_data_device *, uint32_t,
-		struct wl_surface *, wl_fixed_t, wl_fixed_t, struct wl_data_offer *);
-static void datadevleave(void *, struct wl_data_device *);
-static void datadevmotion(void *, struct wl_data_device *, uint32_t,
-		wl_fixed_t x, wl_fixed_t y);
-static void datadevdrop(void *, struct wl_data_device *);
-static void datadevselection(void *, struct wl_data_device *,
-		struct wl_data_offer *);
-static void dataofferoffer(void *, struct wl_data_offer *, const char *);
-static void datasrctarget(void *, struct wl_data_source *, const char *);
-static void datasrcsend(void *, struct wl_data_source *, const char *, int32_t);
-static void datasrccancelled(void *, struct wl_data_source *);
-
-static void selinit(void);
-static void selnormalize(void);
-static inline int selected(int, int);
 static char *getsel(void);
-static void selcopy(uint32_t);
 static void selscroll(int, int);
 static void selsnap(int *, int *, int);
-static int x2col(int);
-static int y2row(int);
 
-static size_t utf8decode(char *, Rune *, size_t);
 static Rune utf8decodebyte(char, size_t *);
-static size_t utf8encode(Rune, char *);
 static char utf8encodebyte(Rune, size_t);
 static char *utf8strchr(char *s, Rune u);
 static size_t utf8validate(Rune *, size_t);
 
 static ssize_t xwrite(int, const char *, size_t);
-static void *xmalloc(size_t);
 static void *xrealloc(void *, size_t);
-static char *xstrdup(char *);
-
-static void usage(void);
-
-static struct wl_registry_listener reglistener = { regglobal, regglobalremove };
-static struct wl_surface_listener surflistener = { surfenter, surfleave };
-static struct wl_callback_listener framelistener = { framedone };
-static struct wl_keyboard_listener kbdlistener =
-	{ kbdkeymap, kbdenter, kbdleave, kbdkey, kbdmodifiers, kbdrepeatinfo };
-static struct wl_pointer_listener ptrlistener =
-	{ ptrenter, ptrleave, ptrmotion, ptrbutton, ptraxis };
-static struct xdg_wm_base_listener wmlistener = { wmping };
-static struct xdg_surface_listener xdgsurflistener = { xdgsurfconfigure };
-static struct xdg_toplevel_listener xdgtoplevellistener =
-	{ xdgtoplevelconfigure, xdgtoplevelclose };
-static struct wl_data_device_listener datadevlistener =
-	{ datadevoffer, datadeventer, datadevleave, datadevmotion, datadevdrop,
-	  datadevselection };
-static struct wl_data_offer_listener dataofferlistener = { dataofferoffer };
-static struct wl_data_source_listener datasrclistener =
-	{ datasrctarget, datasrcsend, datasrccancelled };
 
 /* Globals */
-static DC dc;
-static Wayland wl;
-static WLD wld;
-static Cursor cursor;
-static Term term;
+Wayland wl;
+WLD wld;
 static CSIEscape csiescseq;
 static STREscape strescseq;
-static int cmdfd;
-static pid_t pid;
-static Selection sel;
-static Repeat repeat;
-static bool needdraw = true;
+pid_t pid;
+Selection sel;
+bool needdraw = true;
 static int iofd = 1;
-static char **opt_cmd  = NULL;
-static char *opt_class = NULL;
-static char *opt_embed = NULL;
-static char *opt_font  = NULL;
-static char *opt_io    = NULL;
-static char *opt_line  = NULL;
-static char *opt_name  = NULL;
-static char *opt_title = NULL;
-static int oldbutton   = 3; /* button event on startup: 3 = release */
-static int oldx, oldy;
+char **opt_cmd  = NULL;
+char *opt_class = NULL;
+char *opt_embed = NULL;
+char *opt_font  = NULL;
+char *opt_io    = NULL;
+char *opt_line  = NULL;
+char *opt_name  = NULL;
+char *opt_title = NULL;
 
-static char *usedfont = NULL;
-static double usedfontsize = 0;
-static double defaultfontsize = 0;
+char *usedfont = NULL;
+double usedfontsize = 0;
+double defaultfontsize = 0;
+
+static struct wl_callback_listener framelistener = { framedone };
+static struct wl_data_offer_listener dataofferlistener = { dataofferoffer };
+static struct wl_data_source_listener datasrclistener =
+      { datasrctarget, datasrcsend, datasrccancelled };
+
 
 static uchar utfbyte[UTF_SIZ + 1] = {0x80,    0, 0xC0, 0xE0, 0xF0};
 static uchar utfmask[UTF_SIZ + 1] = {0xC0, 0x80, 0xE0, 0xF0, 0xF8};
 static Rune utfmin[UTF_SIZ + 1] = {       0,    0,  0x80,  0x800,  0x10000};
 static Rune utfmax[UTF_SIZ + 1] = {0x10FFFF, 0x7F, 0x7FF, 0xFFFF, 0x10FFFF};
 
-/* Font Ring Cache */
-enum {
-	FRC_NORMAL,
-	FRC_ITALIC,
-	FRC_BOLD,
-	FRC_ITALICBOLD
-};
-
-typedef struct {
-	struct wld_font *font;
-	int flags;
-	Rune unicodep;
-} Fontcache;
-
-/* Fontcache is an array now. A new font will be appended to the array. */
-static Fontcache frc[16];
-static int frclen = 0;
+/* config.h array lengths */
+size_t colornamelen = LEN(colorname);
+size_t maxcolornamelen = MAX(LEN(colorname), 256);
+size_t mshortcutslen = LEN(mshortcuts);
+size_t ashortcutslen = LEN(ashortcuts);
+size_t shortcutslen = LEN(shortcuts);
+size_t selmaskslen = LEN(selmasks);
+size_t mappedkeyslen = LEN(mappedkeys);
+size_t keylen = LEN(key);
 
 ssize_t
 xwrite(int fd, const char *s, size_t len)
@@ -776,24 +388,6 @@ selinit(void)
 }
 
 int
-x2col(int x)
-{
-	x -= borderpx;
-	x /= wl.cw;
-
-	return LIMIT(x, 0, term.col-1);
-}
-
-int
-y2row(int y)
-{
-	y -= borderpx;
-	y /= wl.ch;
-
-	return LIMIT(y, 0, term.row-1);
-}
-
-int
 tlinelen(int y)
 {
 	int i = term.col;
@@ -921,108 +515,6 @@ selsnap(int *x, int *y, int direction)
 		}
 		break;
 	}
-}
-
-void
-getbuttoninfo(void)
-{
-	int type;
-	uint state = wl.xkb.mods & ~forceselmod;
-
-	sel.alt = IS_SET(MODE_ALTSCREEN);
-
-	sel.oe.x = x2col(wl.px);
-	sel.oe.y = y2row(wl.py);
-	selnormalize();
-
-	sel.type = SEL_REGULAR;
-	for (type = 1; type < LEN(selmasks); ++type) {
-		if (match(selmasks[type], state)) {
-			sel.type = type;
-			break;
-		}
-	}
-}
-
-void
-wlmousereport(int button, bool release, int x, int y)
-{
-	int len;
-	char buf[40];
-
-	if (!IS_SET(MODE_MOUSEX10)) {
-		button += ((wl.xkb.mods & MOD_MASK_SHIFT) ? 4  : 0)
-			+ ((wl.xkb.mods & MOD_MASK_LOGO ) ? 8  : 0)
-			+ ((wl.xkb.mods & MOD_MASK_CTRL ) ? 16 : 0);
-	}
-
-	if (IS_SET(MODE_MOUSESGR)) {
-		len = snprintf(buf, sizeof(buf), "\033[<%d;%d;%d%c",
-				button, x+1, y+1, release ? 'm' : 'M');
-	} else if (x < 223 && y < 223) {
-		len = snprintf(buf, sizeof(buf), "\033[M%c%c%c",
-				32+button, 32+x+1, 32+y+1);
-	} else {
-		return;
-	}
-
-	ttywrite(buf, len);
-}
-
-void
-wlmousereportbutton(uint32_t button, uint32_t state)
-{
-	bool release = state == WL_POINTER_BUTTON_STATE_RELEASED;
-
-	if (!IS_SET(MODE_MOUSESGR) && release) {
-		button = 3;
-	} else {
-		switch (button) {
-		case BTN_LEFT:
-			button = 0;
-			break;
-		case BTN_MIDDLE:
-			button = 1;
-			break;
-		case BTN_RIGHT:
-			button = 2;
-			break;
-		}
-	}
-
-	oldbutton = release ? 3 : button;
-
-	/* don't report release events when in X10 mode */
-	if (IS_SET(MODE_MOUSEX10) && release) {
-		return;
-	}
-
-	wlmousereport(button, release, oldx, oldy);
-}
-
-void
-wlmousereportmotion(wl_fixed_t fx, wl_fixed_t fy)
-{
-	int x = x2col(wl_fixed_to_int(fx)), y = y2row(wl_fixed_to_int(fy));
-
-	if (x == oldx && y == oldy)
-		return;
-	if (!IS_SET(MODE_MOUSEMOTION) && !IS_SET(MODE_MOUSEMANY))
-		return;
-	/* MOUSE_MOTION: no reporting if no button is pressed */
-	if (IS_SET(MODE_MOUSEMOTION) && oldbutton == 3)
-		return;
-
-	oldx = x;
-	oldy = y;
-	wlmousereport(oldbutton + 32, false, x, y);
-}
-
-void
-wlmousereportaxis(uint32_t axis, wl_fixed_t amount)
-{
-	wlmousereport(64 + (axis == AXIS_VERTICAL ? 4 : 6)
-		+ (amount > 0 ? 1 : 0), false, oldx, oldy);
 }
 
 char *
@@ -3065,271 +2557,6 @@ tresize(int col, int row)
 }
 
 void
-wlresize(int col, int row)
-{
-	union wld_object object;
-
-	wl.tw = MAX(1, col * wl.cw);
-	wl.th = MAX(1, row * wl.ch);
-
-	wld.oldbuffer = wld.buffer;
-	wld.buffer = wld_create_buffer(wld.ctx, wl.w, wl.h,
-			WLD_FORMAT_XRGB8888, 0);
-	wld_export(wld.buffer, WLD_WAYLAND_OBJECT_BUFFER, &object);
-	wl.buffer = object.ptr;
-}
-
-uchar
-sixd_to_8bit(int x)
-{
-	return x == 0 ? 0 : 0x37 + 0x28 * x;
-}
-
-int
-wlloadcolor(int i, const char *name, uint32_t *color)
-{
-	if (!name) {
-		if (BETWEEN(i, 16, 255)) { /* 256 color */
-			if (i < 6*6*6+16) { /* same colors as xterm */
-				*color = 0xff << 24 | sixd_to_8bit(((i-16)/36)%6) << 16
-					| sixd_to_8bit(((i-16)/6)%6) << 8
-					| sixd_to_8bit(((i-16)/1)%6);
-			} else { /* greyscale */
-				*color = 0xff << 24 | (0x8 + 0xa * (i-(6*6*6+16))) * 0x10101;
-			}
-			return true;
-		} else
-			name = colorname[i];
-	}
-
-	return wld_lookup_named_color(name, color);
-}
-
-void
-wlloadcols(void)
-{
-	int i;
-
-	for (i = 0; i < LEN(dc.col); i++)
-		if (!wlloadcolor(i, NULL, &dc.col[i])) {
-			if (colorname[i])
-				die("Could not allocate color '%s'\n", colorname[i]);
-			else
-				die("Could not allocate color %d\n", i);
-		}
-}
-
-int
-wlsetcolorname(int x, const char *name)
-{
-	uint32_t color;
-
-	if (!BETWEEN(x, 0, LEN(dc.col)))
-		return 1;
-
-	if (!wlloadcolor(x, name, &color))
-		return 1;
-
-	dc.col[x] = color;
-
-	return 0;
-}
-
-static void wlloadcursor(void)
-{
-	char *names[] = { mouseshape, "xterm", "ibeam", "text" };
-	int i;
-
-	cursor.theme = wl_cursor_theme_load(NULL, 32, wl.shm);
-
-	for (i = 0; !cursor.cursor && i < LEN(names); i++)
-		cursor.cursor = wl_cursor_theme_get_cursor(cursor.theme, names[i]);
-
-	cursor.surface = wl_compositor_create_surface(wl.cmp);
-}
-
-/*
- * Absolute coordinates.
- */
-void
-wlclear(int x1, int y1, int x2, int y2)
-{
-	uint32_t color = dc.col[IS_SET(MODE_REVERSE) ? defaultfg : defaultbg];
-
-	wld_fill_rectangle(wld.renderer, color, x1, y1, x2 - x1, y2 - y1);
-}
-
-int
-wlloadfont(Font *f, FcPattern *pattern)
-{
-	FcPattern *configured;
-	FcPattern *match;
-	FcResult result;
-	struct wld_extents extents;
-	int wantattr, haveattr;
-
-	/*
-	 * Manually configure instead of calling XftMatchFont
-	 * so that we can use the configured pattern for
-	 * "missing glyph" lookups.
-	 */
-	configured = FcPatternDuplicate(pattern);
-	if (!configured)
-		return 1;
-
-	FcConfigSubstitute(NULL, configured, FcMatchPattern);
-	FcDefaultSubstitute(configured);
-
-	match = FcFontMatch(NULL, configured, &result);
-	if (!match) {
-		FcPatternDestroy(configured);
-		return 1;
-	}
-
-	if (!(f->match = wld_font_open_pattern(wld.fontctx, match))) {
-		FcPatternDestroy(configured);
-		FcPatternDestroy(match);
-		return 1;
-	}
-
-	if ((FcPatternGetInteger(pattern, "slant", 0, &wantattr) ==
-	    FcResultMatch)) {
-		/*
-		 * Check if xft was unable to find a font with the appropriate
-		 * slant but gave us one anyway. Try to mitigate.
-		 */
-		if ((FcPatternGetInteger(match, "slant", 0,
-		    &haveattr) != FcResultMatch) || haveattr < wantattr) {
-			f->badslant = 1;
-			fputs("st-wl: font slant does not match\n", stderr);
-		}
-	}
-
-	if ((FcPatternGetInteger(pattern, "weight", 0, &wantattr) ==
-	    FcResultMatch)) {
-		if ((FcPatternGetInteger(match, "weight", 0,
-		    &haveattr) != FcResultMatch) || haveattr != wantattr) {
-			f->badweight = 1;
-			fputs("st-wl: font weight does not match\n", stderr);
-		}
-	}
-
-
-	wld_font_text_extents(f->match, ascii_printable, &extents);
-
-	f->set = NULL;
-	f->pattern = configured;
-
-	f->ascent = f->match->ascent;
-	f->descent = f->match->descent;
-	f->lbearing = 0;
-	f->rbearing = f->match->max_advance;
-
-	f->height = f->ascent + f->descent;
-	f->width = DIVCEIL(extents.advance, strlen(ascii_printable));
-
-	return 0;
-}
-
-void
-wlloadfonts(char *fontstr, double fontsize)
-{
-	FcPattern *pattern;
-	double fontval;
-	float ceilf(float);
-
-	if (fontstr[0] == '-') {
-		/* XXX: need XftXlfdParse equivalent */
-		pattern = NULL;
-	} else {
-		pattern = FcNameParse((FcChar8 *)fontstr);
-	}
-
-	if (!pattern)
-		die("st-wl: can't open font %s\n", fontstr);
-
-	if (fontsize > 1) {
-		FcPatternDel(pattern, FC_PIXEL_SIZE);
-		FcPatternDel(pattern, FC_SIZE);
-		FcPatternAddDouble(pattern, FC_PIXEL_SIZE, (double)fontsize);
-		usedfontsize = fontsize;
-	} else {
-		if (FcPatternGetDouble(pattern, FC_PIXEL_SIZE, 0, &fontval) ==
-				FcResultMatch) {
-			usedfontsize = fontval;
-		} else if (FcPatternGetDouble(pattern, FC_SIZE, 0, &fontval) ==
-				FcResultMatch) {
-			usedfontsize = -1;
-		} else {
-			/*
-			 * Default font size is 12, if none given. This is to
-			 * have a known usedfontsize value.
-			 */
-			FcPatternAddDouble(pattern, FC_PIXEL_SIZE, 12);
-			usedfontsize = 12;
-		}
-		defaultfontsize = usedfontsize;
-	}
-
-	FcConfigSubstitute(0, pattern, FcMatchPattern);
-	FcDefaultSubstitute(pattern);
-
-	if (wlloadfont(&dc.font, pattern))
-		die("st-wl: can't open font %s\n", fontstr);
-
-	if (usedfontsize < 0) {
-		FcPatternGetDouble(dc.font.pattern,
-		                   FC_PIXEL_SIZE, 0, &fontval);
-		usedfontsize = fontval;
-		if (fontsize == 0)
-			defaultfontsize = fontval;
-	}
-
-	/* Setting character width and height. */
-	wl.cw = ceilf(dc.font.width * cwscale);
-	wl.ch = ceilf(dc.font.height * chscale);
-
-	FcPatternDel(pattern, FC_SLANT);
-	FcPatternAddInteger(pattern, FC_SLANT, FC_SLANT_ITALIC);
-	if (wlloadfont(&dc.ifont, pattern))
-		die("st-wl: can't open font %s\n", fontstr);
-
-	FcPatternDel(pattern, FC_WEIGHT);
-	FcPatternAddInteger(pattern, FC_WEIGHT, FC_WEIGHT_BOLD);
-	if (wlloadfont(&dc.ibfont, pattern))
-		die("st-wl: can't open font %s\n", fontstr);
-
-	FcPatternDel(pattern, FC_SLANT);
-	FcPatternAddInteger(pattern, FC_SLANT, FC_SLANT_ROMAN);
-	if (wlloadfont(&dc.bfont, pattern))
-		die("st-wl: can't open font %s\n", fontstr);
-
-	FcPatternDestroy(pattern);
-}
-
-void
-wlunloadfont(Font *f)
-{
-	wld_font_close(f->match);
-	FcPatternDestroy(f->pattern);
-	if (f->set)
-		FcFontSetDestroy(f->set);
-}
-
-void
-wlunloadfonts(void)
-{
-	/* Free the loaded fonts in the font cache.  */
-	while (frclen > 0)
-		wld_font_close(frc[--frclen].font);
-
-	wlunloadfont(&dc.font);
-	wlunloadfont(&dc.bfont);
-	wlunloadfont(&dc.ifont);
-	wlunloadfont(&dc.ibfont);
-}
-
-void
 wlzoom(const Arg *arg)
 {
 	Arg larg;
@@ -3361,440 +2588,6 @@ wlzoomreset(const Arg *arg)
 		larg.f = defaultfontsize;
 		wlzoomabs(&larg);
 	}
-}
-
-void
-wlinit(void)
-{
-	struct wl_registry *registry;
-
-	if (!(wl.dpy = wl_display_connect(NULL)))
-		die("Can't open display\n");
-
-	registry = wl_display_get_registry(wl.dpy);
-	wl_registry_add_listener(registry, &reglistener, NULL);
-	wld.ctx = wld_wayland_create_context(wl.dpy, WLD_ANY);
-	wld.renderer = wld_create_renderer(wld.ctx);
-
-	wl_display_roundtrip(wl.dpy);
-
-	if (!wl.shm)
-		die("Display has no SHM\n");
-	if (!wl.seat)
-		die("Display has no seat\n");
-	if (!wl.datadevmanager)
-		die("Display has no data device manager\n");
-    if (!wl.wm)
-        die("Display has no window manager\n");
-
-    wl.keyboard = wl_seat_get_keyboard(wl.seat);
-	wl_keyboard_add_listener(wl.keyboard, &kbdlistener, NULL);
-	wl.pointer = wl_seat_get_pointer(wl.seat);
-	wl_pointer_add_listener(wl.pointer, &ptrlistener, NULL);
-	wl.datadev = wl_data_device_manager_get_data_device(wl.datadevmanager,
-			wl.seat);
-	wl_data_device_add_listener(wl.datadev, &datadevlistener, NULL);
-
-	/* font */
-	if (!FcInit())
-		die("Could not init fontconfig.\n");
-
-	usedfont = (opt_font == NULL)? font : opt_font;
-	wld.fontctx = wld_font_create_context();
-	wlloadfonts(usedfont, 0);
-
-	wlloadcols();
-	wlloadcursor();
-
-	wl.vis = 0;
-	wl.h = 2 * borderpx + term.row * wl.ch;
-	wl.w = 2 * borderpx + term.col * wl.cw;
-
-	wl.surface = wl_compositor_create_surface(wl.cmp);
-	wl_surface_add_listener(wl.surface, &surflistener, NULL);
-
-	wl.xdgsurface = xdg_wm_base_get_xdg_surface(wl.wm, wl.surface);
-	xdg_surface_add_listener(wl.xdgsurface, &xdgsurflistener, NULL);
-	wl.xdgtoplevel = xdg_surface_get_toplevel(wl.xdgsurface);
-    xdg_toplevel_add_listener(wl.xdgtoplevel, &xdgtoplevellistener, NULL);
-    xdg_toplevel_set_app_id(wl.xdgtoplevel, opt_class ? opt_class : termname);
-
-    wl_surface_commit(wl.surface);
-
-	wl.xkb.ctx = xkb_context_new(0);
-	wlresettitle();
-}
-
-/*
- * TODO: Implement something like XftDrawGlyphFontSpec in wld, and then apply a
- * similar patch to ae1923d27533ff46400d93765e971558201ca1ee
- */
-
-void
-wldraws(char *s, Glyph base, int x, int y, int charlen, int bytelen)
-{
-	int winx = borderpx + x * wl.cw, winy = borderpx + y * wl.ch,
-	    width = charlen * wl.cw, xp, i;
-	int frcflags, charexists;
-	int u8fl, u8fblen, u8cblen, doesexist;
-	char *u8c, *u8fs;
-	Rune unicodep;
-	Font *font = &dc.font;
-	FcResult fcres;
-	FcPattern *fcpattern, *fontpattern;
-	FcFontSet *fcsets[] = { NULL };
-	FcCharSet *fccharset;
-	uint32_t fg, bg, temp;
-	int oneatatime;
-
-	frcflags = FRC_NORMAL;
-
-	/* Fallback on color display for attributes not supported by the font */
-	if (base.mode & ATTR_ITALIC && base.mode & ATTR_BOLD) {
-		if (dc.ibfont.badslant || dc.ibfont.badweight)
-			base.fg = defaultattr;
-		font = &dc.ibfont;
-		frcflags = FRC_ITALICBOLD;
-	} else if (base.mode & ATTR_ITALIC) {
-		if (dc.ifont.badslant)
-			base.fg = defaultattr;
-		font = &dc.ifont;
-		frcflags = FRC_ITALIC;
-	} else if (base.mode & ATTR_BOLD) {
-		if (dc.bfont.badweight)
-			base.fg = defaultattr;
-		font = &dc.ifont;
-		frcflags = FRC_BOLD;
-	}
-
-	if (IS_TRUECOL(base.fg)) {
-		fg = base.fg | 0xff000000;
-	} else {
-		fg = dc.col[base.fg];
-	}
-
-	if (IS_TRUECOL(base.bg)) {
-		bg = base.bg | 0xff000000;
-	} else {
-		bg = dc.col[base.bg];
-	}
-
-	if (base.mode & ATTR_BOLD) {
-		/*
-		 * change basic system colors [0-7]
-		 * to bright system colors [8-15]
-		 */
-		if (BETWEEN(base.fg, 0, 7) && !(base.mode & ATTR_FAINT))
-			fg = dc.col[base.fg + 8];
-
-		if (base.mode & ATTR_ITALIC) {
-			font = &dc.ibfont;
-			frcflags = FRC_ITALICBOLD;
-		} else {
-			font = &dc.bfont;
-			frcflags = FRC_BOLD;
-		}
-	}
-
-	if (IS_SET(MODE_REVERSE)) {
-		if (fg == dc.col[defaultfg]) {
-			fg = dc.col[defaultbg];
-		} else {
-			fg = ~(fg & 0xffffff);
-		}
-
-		if (bg == dc.col[defaultbg]) {
-			bg = dc.col[defaultfg];
-		} else {
-			bg = ~(bg & 0xffffff);
-		}
-	}
-
-	if (base.mode & ATTR_REVERSE) {
-		temp = fg;
-		fg = bg;
-		bg = temp;
-	}
-
-	if (base.mode & ATTR_FAINT && !(base.mode & ATTR_BOLD)) {
-		fg = (fg & (0xff << 24))
-			| ((((fg >> 16) & 0xff) / 2) << 16)
-			| ((((fg >> 8) & 0xff) / 2) << 8)
-			| ((fg & 0xff) / 2);
-	}
-
-	if (base.mode & ATTR_BLINK && term.mode & MODE_BLINK)
-		fg = bg;
-
-	if (base.mode & ATTR_INVISIBLE)
-		fg = bg;
-
-	/* Intelligent cleaning up of the borders. */
-	if (x == 0) {
-		wlclear(0, (y == 0)? 0 : winy, borderpx,
-			((y >= term.row-1)? wl.h : (winy + wl.ch)));
-	}
-	if (x + charlen >= term.col) {
-		wlclear(winx + width, (y == 0)? 0 : winy, wl.w,
-			((y >= term.row-1)? wl.h : (winy + wl.ch)));
-	}
-	if (y == 0)
-		wlclear(winx, 0, winx + width, borderpx);
-	if (y == term.row-1)
-		wlclear(winx, winy + wl.ch, winx + width, wl.h);
-
-	/* Clean up the region we want to draw to. */
-	wld_fill_rectangle(wld.renderer, bg, winx, winy, width, wl.ch);
-
-	for (xp = winx; bytelen > 0;) {
-		/*
-		 * Search for the range in the to be printed string of glyphs
-		 * that are in the main font. Then print that range. If
-		 * some glyph is found that is not in the font, do the
-		 * fallback dance.
-		 */
-		u8fs = s;
-		u8fblen = 0;
-		u8fl = 0;
-		oneatatime = font->width != wl.cw;
-		for (;;) {
-			u8c = s;
-			u8cblen = utf8decode(s, &unicodep, UTF_SIZ);
-			s += u8cblen;
-			bytelen -= u8cblen;
-
-			doesexist = wld_font_ensure_char(font->match, unicodep);
-			if (doesexist) {
-					u8fl++;
-					u8fblen += u8cblen;
-					if (!oneatatime && bytelen > 0)
-							continue;
-			}
-
-			if (u8fl > 0) {
-				wld_draw_text(wld.renderer,
-						font->match, fg, xp,
-						winy + font->ascent,
-						u8fs, u8fblen, NULL);
-				xp += wl.cw * u8fl;
-			}
-			break;
-		}
-		if (doesexist) {
-			if (oneatatime)
-				continue;
-			break;
-		}
-
-		/* Search the font cache. */
-		for (i = 0; i < frclen; i++) {
-			charexists = wld_font_ensure_char(frc[i].font, unicodep);
-			/* Everything correct. */
-			if (charexists && frc[i].flags == frcflags)
-				break;
-			/* We got a default font for a not found glyph. */
-			if (!charexists && frc[i].flags == frcflags \
-					&& frc[i].unicodep == unicodep) {
-				break;
-			}
-		}
-
-		/* Nothing was found. */
-		if (i >= frclen) {
-			if (!font->set)
-				font->set = FcFontSort(0, font->pattern,
-				                       1, 0, &fcres);
-			fcsets[0] = font->set;
-
-			/*
-			 * Nothing was found in the cache. Now use
-			 * some dozen of Fontconfig calls to get the
-			 * font for one single character.
-			 *
-			 * Xft and fontconfig are design failures.
-			 */
-			fcpattern = FcPatternDuplicate(font->pattern);
-			fccharset = FcCharSetCreate();
-
-			FcCharSetAddChar(fccharset, unicodep);
-			FcPatternAddCharSet(fcpattern, FC_CHARSET,
-					fccharset);
-			FcPatternAddBool(fcpattern, FC_SCALABLE, 1);
-
-			FcConfigSubstitute(0, fcpattern,
-					FcMatchPattern);
-			FcDefaultSubstitute(fcpattern);
-
-			fontpattern = FcFontSetMatch(0, fcsets, 1,
-					fcpattern, &fcres);
-
-			/*
-			 * Overwrite or create the new cache entry.
-			 */
-			if (frclen >= LEN(frc)) {
-				frclen = LEN(frc) - 1;
-				wld_font_close(frc[frclen].font);
-				frc[frclen].unicodep = 0;
-			}
-
-			frc[frclen].font = wld_font_open_pattern(wld.fontctx,
-					fontpattern);
-			frc[frclen].flags = frcflags;
-			frc[frclen].unicodep = unicodep;
-
-			i = frclen;
-			frclen++;
-
-			FcPatternDestroy(fcpattern);
-			FcCharSetDestroy(fccharset);
-		}
-
-		wld_draw_text(wld.renderer, frc[i].font, fg,
-				xp, winy + frc[i].font->ascent,
-				u8c, u8cblen, NULL);
-
-		xp += wl.cw * wcwidth(unicodep);
-	}
-
-	if (base.mode & ATTR_UNDERLINE) {
-		wld_fill_rectangle(wld.renderer, fg, winx, winy + font->ascent + 1,
-				width, 1);
-	}
-
-	if (base.mode & ATTR_STRUCK) {
-		wld_fill_rectangle(wld.renderer, fg, winx, winy + 2 * font->ascent / 3,
-				width, 1);
-	}
-}
-
-void
-wldrawglyph(Glyph g, int x, int y)
-{
-	static char buf[UTF_SIZ];
-	size_t len = utf8encode(g.u, buf);
-	int width = g.mode & ATTR_WIDE ? 2 : 1;
-
-	wldraws(buf, g, x, y, width, len);
-}
-
-void
-wldrawcursor(void)
-{
-	static int oldx = 0, oldy = 0;
-	int curx;
-	Glyph g = {' ', ATTR_NULL, defaultbg, defaultcs}, og;
-	int ena_sel = sel.ob.x != -1 && sel.alt == IS_SET(MODE_ALTSCREEN);
-	uint32_t drawcol;
-
-	LIMIT(oldx, 0, term.col-1);
-	LIMIT(oldy, 0, term.row-1);
-
-	curx = term.c.x;
-
-	/* adjust position if in dummy */
-	if (term.line[oldy][oldx].mode & ATTR_WDUMMY)
-		oldx--;
-	if (term.line[term.c.y][curx].mode & ATTR_WDUMMY)
-		curx--;
-
-	/* remove the old cursor */
-	og = term.line[oldy][oldx];
-	if (ena_sel && selected(oldx, oldy))
-		og.mode ^= ATTR_REVERSE;
-	wldrawglyph(og, oldx, oldy);
-	if (oldx != curx || oldy != term.c.y) {
-		wl_surface_damage(wl.surface, borderpx + oldx * wl.cw,
-				borderpx + oldy * wl.ch, wl.cw, wl.ch);
-	}
-
-	g.u = term.line[term.c.y][term.c.x].u;
-
-	/*
-	 * Select the right color for the right mode.
-	 */
-	if (IS_SET(MODE_REVERSE)) {
-		g.mode |= ATTR_REVERSE;
-		g.bg = defaultfg;
-		if (ena_sel && selected(term.c.x, term.c.y)) {
-			drawcol = dc.col[defaultcs];
-			g.fg = defaultrcs;
-		} else {
-			drawcol = dc.col[defaultrcs];
-			g.fg = defaultcs;
-		}
-	} else {
-		if (ena_sel && selected(term.c.x, term.c.y)) {
-			drawcol = dc.col[defaultrcs];
-			g.fg = defaultfg;
-			g.bg = defaultrcs;
-		} else {
-			drawcol = dc.col[defaultcs];
-		}
-	}
-
-	if (IS_SET(MODE_HIDE))
-		return;
-
-	/* draw the new one */
-	if (wl.state & WIN_FOCUSED) {
-		switch (wl.cursor) {
-		case 7: /* st-wl extension: snowman */
-			utf8decode("☃", &g.u, UTF_SIZ);
-		case 0: /* Blinking Block */
-		case 1: /* Blinking Block (Default) */
-		case 2: /* Steady Block */
-			g.mode |= term.line[term.c.y][curx].mode & ATTR_WIDE;
-			wldrawglyph(g, term.c.x, term.c.y);
-			break;
-		case 3: /* Blinking Underline */
-		case 4: /* Steady Underline */
-			wld_fill_rectangle(wld.renderer, drawcol,
-					borderpx + curx * wl.cw,
-					borderpx + (term.c.y + 1) * wl.ch - \
-						cursorthickness,
-					wl.cw, cursorthickness);
-			break;
-		case 5: /* Blinking bar */
-		case 6: /* Steady bar */
-			wld_fill_rectangle(wld.renderer, drawcol,
-					borderpx + curx * wl.cw,
-					borderpx + term.c.y * wl.ch,
-					cursorthickness, wl.ch);
-			break;
-		}
-	} else {
-		wld_fill_rectangle(wld.renderer, drawcol,
-				borderpx + curx * wl.cw,
-				borderpx + term.c.y * wl.ch,
-				wl.cw - 1, 1);
-		wld_fill_rectangle(wld.renderer, drawcol,
-				borderpx + curx * wl.cw,
-				borderpx + term.c.y * wl.ch,
-				1, wl.ch - 1);
-		wld_fill_rectangle(wld.renderer, drawcol,
-				borderpx + (curx + 1) * wl.cw - 1,
-				borderpx + term.c.y * wl.ch,
-				1, wl.ch - 1);
-		wld_fill_rectangle(wld.renderer, drawcol,
-				borderpx + curx * wl.cw,
-				borderpx + (term.c.y + 1) * wl.ch - 1,
-				wl.cw, 1);
-	}
-	wl_surface_damage(wl.surface, borderpx + curx * wl.cw,
-			borderpx + term.c.y * wl.ch, wl.cw, wl.ch);
-	oldx = curx, oldy = term.c.y;
-}
-
-void
-wlsettitle(char *title)
-{
-	xdg_toplevel_set_title(wl.xdgtoplevel, title);
-}
-
-void
-wlresettitle(void)
-{
-	wlsettitle(opt_title ? opt_title : "st-wl");
 }
 
 void
@@ -3878,56 +2671,10 @@ wlseturgency(int add)
 	/* XXX: no urgency equivalent yet in wayland */
 }
 
-int
-match(uint mask, uint state)
-{
-	return mask == MOD_MASK_ANY || mask == (state & ~(ignoremod));
-}
-
 void
 numlock(const Arg *dummy)
 {
 	term.numlock ^= 1;
-}
-
-char*
-kmap(xkb_keysym_t k, uint state)
-{
-	Key *kp;
-	int i;
-
-	/* Check for mapped keys out of X11 function keys. */
-	for (i = 0; i < LEN(mappedkeys); i++) {
-		if (mappedkeys[i] == k)
-			break;
-	}
-	if (i == LEN(mappedkeys)) {
-		if ((k & 0xFFFF) < 0xFD00)
-			return NULL;
-	}
-
-	for (kp = key; kp < key + LEN(key); kp++) {
-		if (kp->k != k)
-			continue;
-
-		if (!match(kp->mask, state))
-			continue;
-
-		if (IS_SET(MODE_APPKEYPAD) ? kp->appkey < 0 : kp->appkey > 0)
-			continue;
-		if (term.numlock && kp->appkey == 2)
-			continue;
-
-		if (IS_SET(MODE_APPCURSOR) ? kp->appcursor < 0 : kp->appcursor > 0)
-			continue;
-
-		if (IS_SET(MODE_CRLF) ? kp->crlf < 0 : kp->crlf > 0)
-			continue;
-
-		return kp->s;
-	}
-
-	return NULL;
 }
 
 void
@@ -3945,388 +2692,6 @@ cresize(int width, int height)
 
 	tresize(col, row);
 	wlresize(col, row);
-}
-
-void
-regglobal(void *data, struct wl_registry *registry, uint32_t name,
-          const char *interface, uint32_t version)
-{
-	if (strcmp(interface, "wl_compositor") == 0) {
-		wl.cmp = wl_registry_bind(registry, name, &wl_compositor_interface, 3);
-	} else if (strcmp(interface, "xdg_wm_base") == 0) {
-		wl.wm = wl_registry_bind(registry, name, &xdg_wm_base_interface, 1);
-		xdg_wm_base_add_listener(wl.wm, &wmlistener, NULL);
-	} else if (strcmp(interface, "wl_shm") == 0) {
-		wl.shm = wl_registry_bind(registry, name, &wl_shm_interface, 1);
-	} else if (strcmp(interface, "wl_seat") == 0) {
-		wl.seat = wl_registry_bind(registry, name, &wl_seat_interface, 4);
-	} else if (strcmp(interface, "wl_data_device_manager") == 0) {
-		wl.datadevmanager = wl_registry_bind(registry, name,
-				&wl_data_device_manager_interface, 1);
-	} else if (strcmp(interface, "wl_output") == 0) {
-		/* bind to outputs so we can get surface enter events */
-		wl_registry_bind(registry, name, &wl_output_interface, 2);
-	}
-}
-
-void
-regglobalremove(void *data, struct wl_registry *registry, uint32_t name)
-{
-}
-
-void
-surfenter(void *data, struct wl_surface *surface, struct wl_output *output)
-{
-	wl.vis++;
-	if (!(wl.state & WIN_VISIBLE))
-		wl.state |= WIN_VISIBLE;
-}
-
-void
-surfleave(void *data, struct wl_surface *surface, struct wl_output *output)
-{
-	if (--wl.vis == 0)
-		wl.state &= ~WIN_VISIBLE;
-}
-
-void
-framedone(void *data, struct wl_callback *callback, uint32_t msecs)
-{
-	wl_callback_destroy(callback);
-	wl.framecb = NULL;
-	if (needdraw && wl.state & WIN_VISIBLE) {
-		draw();
-	}
-}
-
-void
-kbdkeymap(void *data, struct wl_keyboard *keyboard, uint32_t format, int32_t fd,
-          uint32_t size)
-{
-	char *string;
-
-	if (format != WL_KEYBOARD_KEYMAP_FORMAT_XKB_V1) {
-		close(fd);
-		return;
-	}
-
-	string = mmap(NULL, size, PROT_READ, MAP_SHARED, fd, 0);
-
-	if (string == MAP_FAILED) {
-		close(fd);
-		return;
-	}
-
-	wl.xkb.keymap = xkb_keymap_new_from_string(wl.xkb.ctx, string,
-			XKB_KEYMAP_FORMAT_TEXT_V1, 0);
-	munmap(string, size);
-	close(fd);
-	wl.xkb.state = xkb_state_new(wl.xkb.keymap);
-
-	wl.xkb.ctrl = xkb_keymap_mod_get_index(wl.xkb.keymap, XKB_MOD_NAME_CTRL);
-	wl.xkb.alt = xkb_keymap_mod_get_index(wl.xkb.keymap, XKB_MOD_NAME_ALT);
-	wl.xkb.shift = xkb_keymap_mod_get_index(wl.xkb.keymap, XKB_MOD_NAME_SHIFT);
-	wl.xkb.logo = xkb_keymap_mod_get_index(wl.xkb.keymap, XKB_MOD_NAME_LOGO);
-
-	wl.xkb.mods = 0;
-}
-
-void
-kbdenter(void *data, struct wl_keyboard *keyboard, uint32_t serial,
-         struct wl_surface *surface, struct wl_array *keys)
-{
-	wl.state |= WIN_FOCUSED;
-	if (IS_SET(MODE_FOCUS))
-		ttywrite("\033[I", 3);
-	/* need to redraw the cursor */
-	needdraw = true;
-}
-
-void
-kbdleave(void *data, struct wl_keyboard *keyboard, uint32_t serial,
-	 struct wl_surface *surface)
-{
-	/* selection offers are invalidated when we lose keyboard focus */
-	wl.seloffer = NULL;
-	wl.state &= ~WIN_FOCUSED;
-	if (IS_SET(MODE_FOCUS))
-		ttywrite("\033[O", 3);
-	/* need to redraw the cursor */
-	needdraw = true;
-	/* disable key repeat */
-	repeat.len = 0;
-}
-
-void
-kbdkey(void *data, struct wl_keyboard *keyboard, uint32_t serial, uint32_t time,
-       uint32_t key, uint32_t state)
-{
-	xkb_keysym_t ksym;
-	char buf[32], *str;
-	int len;
-	Rune c;
-	Shortcut *bp;
-
-	if (IS_SET(MODE_KBDLOCK))
-		return;
-
-	if (state == WL_KEYBOARD_KEY_STATE_RELEASED) {
-		if (repeat.key == key)
-			repeat.len = 0;
-		return;
-	}
-
-	ksym = xkb_state_key_get_one_sym(wl.xkb.state, key + 8);
-	len = xkb_keysym_to_utf8(ksym, buf, sizeof buf);
-	if (len > 0)
-	    --len;
-
-	/* 1. shortcuts */
-	for (bp = shortcuts; bp < shortcuts + LEN(shortcuts); bp++) {
-		if (ksym == bp->keysym && match(bp->mod, wl.xkb.mods)) {
-			bp->func(&(bp->arg));
-			return;
-		}
-	}
-
-	/* 2. custom keys from config.h */
-	if ((str = kmap(ksym, wl.xkb.mods))) {
-		len = strlen(str);
-		goto send;
-	}
-
-	/* 3. composed string from input method */
-	if (len == 0)
-		return;
-	if (len == 1 && wl.xkb.mods & MOD_MASK_ALT) {
-		if (IS_SET(MODE_8BIT)) {
-			if (*buf < 0177) {
-				c = *buf | 0x80;
-				len = utf8encode(c, buf);
-			}
-		} else {
-			buf[1] = buf[0];
-			buf[0] = '\033';
-			len = 2;
-		}
-	}
-	/* convert character to control character */
-	else if (len == 1 && wl.xkb.mods & MOD_MASK_CTRL) {
-		if ((*buf >= '@' && *buf < '\177') || *buf == ' ')
-			*buf &= 0x1F;
-		else if (*buf == '2') *buf = '\000';
-		else if (*buf >= '3' && *buf <= '7')
-			*buf -= ('3' - '\033');
-		else if (*buf == '8') *buf = '\177';
-		else if (*buf == '/') *buf = '_' & 0x1F;
-	}
-
-	str = buf;
-
-send:
-	memcpy(repeat.str, str, len);
-	repeat.key = key;
-	repeat.len = len;
-	repeat.started = false;
-	clock_gettime(CLOCK_MONOTONIC, &repeat.last);
-	ttysend(str, len);
-}
-
-void
-kbdmodifiers(void *data, struct wl_keyboard *keyboard, uint32_t serial,
-             uint32_t dep, uint32_t lat, uint32_t lck, uint32_t group)
-{
-	xkb_mod_mask_t mod_mask;
-
-	xkb_state_update_mask(wl.xkb.state, dep, lat, lck, group, 0, 0);
-
-	mod_mask = xkb_state_serialize_mods(wl.xkb.state, XKB_STATE_MODS_EFFECTIVE);
-	wl.xkb.mods = 0;
-
-	if (mod_mask & (1 << wl.xkb.ctrl))
-		wl.xkb.mods |= MOD_MASK_CTRL;
-	if (mod_mask & (1 << wl.xkb.alt))
-		wl.xkb.mods |= MOD_MASK_ALT;
-	if (mod_mask & (1 << wl.xkb.shift))
-		wl.xkb.mods |= MOD_MASK_SHIFT;
-	if (mod_mask & (1 << wl.xkb.logo))
-		wl.xkb.mods |= MOD_MASK_LOGO;
-}
-
-void
-kbdrepeatinfo(void *data, struct wl_keyboard *keyboard, int32_t rate,
-              int32_t delay)
-{
-	keyrepeatdelay = delay;
-	keyrepeatinterval = 1000 / rate;
-}
-
-void
-ptrenter(void *data, struct wl_pointer *pointer, uint32_t serial,
-         struct wl_surface *surface, wl_fixed_t x, wl_fixed_t y)
-{
-	struct wl_cursor_image *img = cursor.cursor->images[0];
-	struct wl_buffer *buffer;
-
-	wl_pointer_set_cursor(pointer, serial, cursor.surface,
-			img->hotspot_x, img->hotspot_y);
-	buffer = wl_cursor_image_get_buffer(img);
-	wl_surface_attach(cursor.surface, buffer, 0, 0);
-	wl_surface_damage(cursor.surface, 0, 0, img->width, img->height);
-	wl_surface_commit(cursor.surface);
-}
-
-void
-ptrleave(void *data, struct wl_pointer *pointer, uint32_t serial,
-         struct wl_surface *surface)
-{
-}
-
-void
-ptrmotion(void *data, struct wl_pointer * pointer, uint32_t serial,
-          wl_fixed_t x, wl_fixed_t y)
-{
-	int oldey, oldex, oldsby, oldsey;
-
-	if (IS_SET(MODE_MOUSE)) {
-		wlmousereportmotion(x, y);
-		return;
-	}
-
-	wl.px = wl_fixed_to_int(x);
-	wl.py = wl_fixed_to_int(y);
-
-	if (!sel.mode)
-		return;
-
-	sel.mode = SEL_READY;
-	oldey = sel.oe.y;
-	oldex = sel.oe.x;
-	oldsby = sel.nb.y;
-	oldsey = sel.ne.y;
-	getbuttoninfo();
-
-	if (oldey != sel.oe.y || oldex != sel.oe.x)
-		tsetdirt(MIN(sel.nb.y, oldsby), MAX(sel.ne.y, oldsey));
-}
-
-void
-ptrbutton(void * data, struct wl_pointer * pointer, uint32_t serial,
-          uint32_t time, uint32_t button, uint32_t state)
-{
-	MouseShortcut *ms;
-
-	if (IS_SET(MODE_MOUSE) && !(wl.xkb.mods & forceselmod)) {
-		wlmousereportbutton(button, state);
-		return;
-	}
-
-	switch (state) {
-	case WL_POINTER_BUTTON_STATE_RELEASED:
-		if (button == BTN_MIDDLE) {
-			selpaste(NULL);
-		} else if (button == BTN_LEFT) {
-			if (sel.mode == SEL_READY) {
-				getbuttoninfo();
-				selcopy(serial);
-			} else
-				selclear();
-			sel.mode = SEL_IDLE;
-			tsetdirt(sel.nb.y, sel.ne.y);
-		}
-		break;
-
-	case WL_POINTER_BUTTON_STATE_PRESSED:
-		for (ms = mshortcuts; ms < mshortcuts + LEN(mshortcuts); ms++) {
-			if (button == ms->b && match(ms->mask, wl.xkb.mods)) {
-				ttysend(ms->s, strlen(ms->s));
-				return;
-			}
-		}
-
-		if (button == BTN_LEFT) {
-			/* Clear previous selection, logically and visually. */
-			selclear();
-			sel.mode = SEL_EMPTY;
-			sel.type = SEL_REGULAR;
-			sel.oe.x = sel.ob.x = x2col(wl.px);
-			sel.oe.y = sel.ob.y = y2row(wl.py);
-
-			/*
-			 * If the user clicks below predefined timeouts
-			 * specific snapping behaviour is exposed.
-			 */
-			if (time - sel.tclick2 <= tripleclicktimeout) {
-				sel.snap = SNAP_LINE;
-			} else if (time - sel.tclick1 <= doubleclicktimeout) {
-				sel.snap = SNAP_WORD;
-			} else {
-				sel.snap = 0;
-			}
-			selnormalize();
-
-			if (sel.snap != 0)
-				sel.mode = SEL_READY;
-			tsetdirt(sel.nb.y, sel.ne.y);
-			sel.tclick2 = sel.tclick1;
-			sel.tclick1 = time;
-		}
-		break;
-	}
-}
-
-void
-ptraxis(void * data, struct wl_pointer * pointer, uint32_t time, uint32_t axis,
-        wl_fixed_t value)
-{
-	Axiskey *ak;
-	int dir = value > 0 ? +1 : -1;
-
-	if (IS_SET(MODE_MOUSE) && !(wl.xkb.mods & forceselmod)) {
-		wlmousereportaxis(axis, value);
-		return;
-	}
-
-	for (ak = ashortcuts; ak < ashortcuts + LEN(ashortcuts); ak++) {
-		if (axis == ak->axis && dir == ak->dir
-				&& match(ak->mask, wl.xkb.mods)) {
-			ttysend(ak->s, strlen(ak->s));
-			return;
-		}
-	}
-}
-
-void
-wmping(void *data, struct xdg_wm_base *wm, uint32_t serial)
-{
-	xdg_wm_base_pong(wm, serial);
-}
-
-void
-xdgsurfconfigure(void *data, struct xdg_surface *surf, uint32_t serial)
-{
-	xdg_surface_ack_configure(surf, serial);
-}
-
-void
-xdgtoplevelconfigure(void *data, struct xdg_toplevel *toplevel,
-                     int32_t w, int32_t h, struct wl_array *states)
-{
-	if (w == wl.w && h == wl.h)
-		return;
-	cresize(w, h);
-	if (wl.configured)
-		ttyresize();
-	else
-		wl.configured = true;
-}
-
-void
-xdgtoplevelclose(void *data, struct xdg_toplevel *toplevel)
-{
-	/* Send SIGHUP to shell */
-	kill(pid, SIGHUP);
-	exit(0);
 }
 
 void
@@ -4407,96 +2772,6 @@ datasrccancelled(void *data, struct wl_data_source *source)
 }
 
 void
-run(void)
-{
-	fd_set rfd;
-	int wlfd = wl_display_get_fd(wl.dpy), blinkset = 0;
-	struct timespec drawtimeout, *tv = NULL, now, last, lastblink;
-	ulong msecs;
-
-	/* Look for initial configure. */
-	wl_display_roundtrip(wl.dpy);
-	if (!wl.configured)
-		cresize(wl.w, wl.h);
-	ttynew();
-	ttyresize();
-	draw();
-
-	clock_gettime(CLOCK_MONOTONIC, &last);
-	lastblink = last;
-
-	for (;;) {
-		FD_ZERO(&rfd);
-		FD_SET(cmdfd, &rfd);
-		FD_SET(wlfd, &rfd);
-
-		if (pselect(MAX(wlfd, cmdfd)+1, &rfd, NULL, NULL, tv, NULL) < 0) {
-			if (errno == EINTR)
-				continue;
-			die("select failed: %s\n", strerror(errno));
-		}
-
-		if (FD_ISSET(cmdfd, &rfd)) {
-			ttyread();
-			if (blinktimeout) {
-				blinkset = tattrset(ATTR_BLINK);
-				if (!blinkset)
-					MODBIT(term.mode, 0, MODE_BLINK);
-			}
-		}
-
-		if (FD_ISSET(wlfd, &rfd)) {
-			if (wl_display_dispatch(wl.dpy) == -1)
-				die("Connection error\n");
-		}
-
-		clock_gettime(CLOCK_MONOTONIC, &now);
-		msecs = -1;
-
-		if (blinkset && blinktimeout) {
-			if (TIMEDIFF(now, lastblink) >= blinktimeout) {
-				tsetdirtattr(ATTR_BLINK);
-				term.mode ^= MODE_BLINK;
-				lastblink = now;
-			} else {
-				msecs = MIN(msecs, blinktimeout - \
-						TIMEDIFF(now, lastblink));
-			}
-		}
-		if (repeat.len > 0) {
-			if (TIMEDIFF(now, repeat.last) >= \
-				(repeat.started ? keyrepeatinterval : \
-					keyrepeatdelay)) {
-				repeat.started = true;
-				repeat.last = now;
-				ttysend(repeat.str, repeat.len);
-			} else {
-				msecs = MIN(msecs, (repeat.started ? \
-					keyrepeatinterval : keyrepeatdelay) - \
-					TIMEDIFF(now, repeat.last));
-			}
-		}
-
-		if (needdraw && wl.state & WIN_VISIBLE) {
-			if (!wl.framecb) {
-				draw();
-			}
-		}
-
-		if (msecs == -1) {
-			tv = NULL;
-		} else {
-			drawtimeout.tv_nsec = 1E6 * msecs;
-			drawtimeout.tv_sec = 0;
-			tv = &drawtimeout;
-		}
-
-		wl_display_dispatch_pending(wl.dpy);
-		wl_display_flush(wl.dpy);
-	}
-}
-
-void
 usage(void)
 {
 	die("usage: %s [-aiv] [-c class] [-f font] [-g geometry]"
@@ -4508,62 +2783,3 @@ usage(void)
 	    "          [-T title] [-t title] [-w windowid] -l line"
 	    " [stty_args ...]\n", argv0, argv0);
 }
-
-int
-main(int argc, char *argv[])
-{
-	wl.cursor = cursorshape;
-
-	ARGBEGIN {
-	case 'a':
-		allowaltscreen = 0;
-		break;
-	case 'c':
-		opt_class = EARGF(usage());
-		break;
-	case 'e':
-		if (argc > 0)
-			--argc, ++argv;
-		goto run;
-	case 'f':
-		opt_font = EARGF(usage());
-		break;
-	case 'o':
-		opt_io = EARGF(usage());
-		break;
-	case 'l':
-		opt_line = EARGF(usage());
-		break;
-	case 'n':
-		opt_name = EARGF(usage());
-		break;
-	case 't':
-	case 'T':
-		opt_title = EARGF(usage());
-		break;
-	case 'w':
-		opt_embed = EARGF(usage());
-		break;
-	case 'v':
-		die("%s " VERSION " (c) 2010-2016 st-wl engineers\n", argv0);
-		break;
-	default:
-		usage();
-	} ARGEND;
-
-run:
-	if (argc > 0) {
-		/* eat all remaining arguments */
-		opt_cmd = argv;
-		if (!opt_title && !opt_line)
-			opt_title = basename(xstrdup(argv[0]));
-	}
-	setlocale(LC_CTYPE, "");
-	tnew(MAX(cols, 1), MAX(rows, 1));
-	wlinit();
-	selinit();
-	run();
-
-	return 0;
-}
-
